+++ tests/comms/test_room.py
@@ -1,7 +1,7 @@
 from unittest.mock import Mock
 
 from cibo.models.message import Message, MessageRoute
-from tests.conftest import CommsFactory
+from tests.comms.conftest import CommsFactory
 
 
 class TestCommsRoom(CommsFactory):
