+++ tests/comms/test_server.py
@@ -1,5 +1,5 @@
 from cibo.models.message import Message, MessageRoute
-from tests.conftest import CommsFactory
+from tests.comms.conftest import CommsFactory
 
 
 class TestCommsServer(CommsFactory):
