+++ tests/conftest.py
@@ -6,12 +6,6 @@ from peewee import SqliteDatabase
 from pytest import fixture
 
 from cibo.actions.commands._processor_ import CommandProcessor
-from cibo.comms._processor_ import CommsProcessor
-from cibo.comms.private import Private as CommsPrivate
-from cibo.comms.region import Region as CommsRegion
-from cibo.comms.room import Room as CommsRoom
-from cibo.comms.sector import Sector as CommsSector
-from cibo.comms.server import Server as CommsServer
 from cibo.entities.world import World
 from cibo.models.client import Client, ClientLoginState
 from cibo.models.data.item import Item as ItemData
@@ -165,19 +159,6 @@ class MessageFactory:
             "highlight": False,
             "terminal_width": 76,
         }
-        yield
-
-
-class CommsFactory(BaseFactory, ClientFactory, WorldFactory):
-    @fixture(autouse=True)
-    def fixture_comms(self, _fixture_mock_clients):
-        self.telnet.get_connected_clients.return_value = [self.mock_clients[0]]
-        self.comms = CommsProcessor(self.telnet, self.world)
-        self.private = CommsPrivate(self.telnet, self.world)
-        self.room = CommsRoom(self.telnet, self.world)
-        self.sector = CommsSector(self.telnet, self.world)
-        self.region = CommsRegion(self.telnet, self.world)
-        self.server = CommsServer(self.telnet, self.world)
         yield
